+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Typescript types
+- `e.target.dirty` and `e.target.values` not being updated in `inform-input` and `inform-change` events.
 
 ## [1.8.2] - 2022-10-07